refactor: simplify QueryHandler class
BREAKING CHANGE: Simplify QueryHandler class by splitting out into new /
existing classes. The QueryHandler class now has instance variables for
classes (named handlers) that are used to make main app queries
