+++ variation/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.4.0a7"
+__version__ = "0.5.0a0"
